+++ discovery_engine_core/core/src/storage/utils.rs
@@ -12,11 +12,9 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use std::borrow::Cow;
+//! Module containing non-database specific sqlx utilities.
 
 use sqlx::{Database, Encode, QueryBuilder, Type};
-
-use crate::storage::Error;
 
 pub(super) trait SqlxPushTupleExt<'args, DB: Database> {
     fn push_tuple<I>(&mut self, values: I) -> &mut Self
@@ -40,75 +38,5 @@ where
         }
         separated.push_unseparated(")");
         self
-    }
-}
-
-pub(crate) trait SqlxSqliteResultExt<T> {
-    /// In case of a foreign key violation return given error.
-    ///
-    /// Converts other `sqlx::Error`s to `storage::Error`s.
-    fn on_fk_violation(self, error: Error) -> Result<T, Error>;
-
-    /// In case of a row not found error return given error.
-    ///
-    /// Converts other `sqlx::Error`s to `storage::Error`s.
-    fn on_row_not_found(self, error: Error) -> Result<T, Error>;
-}
-
-impl<T> SqlxSqliteResultExt<T> for Result<T, sqlx::Error> {
-    fn on_fk_violation(self, error: Error) -> Result<T, Error> {
-        if let Err(sqlx::Error::Database(db_err)) = &self {
-            if db_err.code() == Some(Cow::Borrowed("787")) {
-                return Err(error);
-            }
-        }
-        self.map_err(Into::into)
-    }
-
-    fn on_row_not_found(self, error: Error) -> Result<T, Error> {
-        if let Err(sqlx::Error::RowNotFound) = &self {
-            Err(error)
-        } else {
-            self.map_err(Into::into)
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{document, storage::sqlite::SqliteStorage};
-
-    use super::*;
-
-    #[tokio::test]
-    async fn test_fk_violation_is_invalid_document() {
-        let storage = SqliteStorage::connect("sqlite::memory:").await.unwrap();
-
-        sqlx::query("CREATE TABLE Foo(x INTEGER PRIMARY KEY);")
-            .execute(&storage.pool)
-            .await
-            .unwrap();
-
-        sqlx::query("CREATE TABLE Bar(x INTEGER PRIMARY KEY REFERENCES Foo(x));")
-            .execute(&storage.pool)
-            .await
-            .unwrap();
-
-        let document_id = document::Id::new();
-
-        let res = sqlx::query("INSERT INTO Bar(x) VALUES (?);")
-            .bind(10u32)
-            .execute(&storage.pool)
-            .await
-            .on_fk_violation(Error::NoDocument(document_id));
-
-        assert!(matches!(res, Err(Error::NoDocument(id)) if id == document_id));
-
-        let res = sqlx::query("malformed;")
-            .execute(&storage.pool)
-            .await
-            .on_fk_violation(Error::NoDocument(document_id));
-
-        assert!(!matches!(res, Err(Error::NoDocument(_))));
     }
 }
